refactor UI trace payload helpers
